tests: migrate tests from tests/lora
