+++ tests/pipelines/cogvideo/test_cogvideox.py
@@ -32,6 +32,8 @@ from ..testing_utils import (
     BasePipelineTesterConfig,
     FasterCacheTesterMixin,
     FirstBlockCacheTesterMixin,
+    LoraMemoryTesterMixin,
+    LoraTesterMixin,
     MemoryTesterMixin,
     PipelineTesterMixin,
     PyramidAttentionBroadcastTesterMixin,
@@ -234,6 +236,21 @@ class TestCogVideoXPipelineFasterCache(CogVideoXPipelineTesterConfig, FasterCach
 
 class TestCogVideoXPipelineFirstBlockCache(CogVideoXPipelineTesterConfig, FirstBlockCacheTesterMixin):
     pass
+
+
+class TestCogVideoXPipelineLoRA(CogVideoXPipelineTesterConfig, LoraTesterMixin):
+    """LoRA tests for the CogVideoX pipeline."""
+
+
+class TestCogVideoXPipelineLoRAMemory(CogVideoXPipelineTesterConfig, LoraMemoryTesterMixin):
+    """LoRA x memory-optimization tests for the CogVideoX pipeline."""
+
+    # `(leaf_level, True)` is left out on purpose, see
+    # https://github.com/huggingface/diffusers/pull/11804#issuecomment-3013325338
+    @pytest.mark.parametrize("offload_type,use_stream", [("block_level", True), ("leaf_level", False)])
+    @require_torch_accelerator
+    def test_group_offloading_inference_denoiser(self, tmp_path, offload_type, use_stream):
+        super().test_group_offloading_inference_denoiser(tmp_path, offload_type, use_stream)
 
 
 @nightly
