+++ tests/pipelines/flux2/test_pipeline_flux2.py
@@ -11,6 +11,8 @@ from diffusers import (
 from ...testing_utils import assert_tensors_close, torch_device
 from ..testing_utils import (
     BasePipelineTesterConfig,
+    LoraMemoryTesterMixin,
+    LoraTesterMixin,
     MemoryTesterMixin,
     PipelineTesterMixin,
     check_qkv_fused_layers_exist,
@@ -187,3 +189,16 @@ class TestFlux2Pipeline(Flux2PipelineTesterConfig, PipelineTesterMixin):
 
 class TestFlux2PipelineMemory(Flux2PipelineTesterConfig, MemoryTesterMixin):
     """Memory optimization tests (CPU offload, group offload, layerwise casting) for the Flux2 pipeline."""
+
+
+class TestFlux2PipelineLoRA(Flux2PipelineTesterConfig, LoraTesterMixin):
+    """LoRA tests for the Flux2 pipeline."""
+
+    # Flux2 fuses the QKV and MLP input projections into a single `to_qkv_mlp_proj` linear.
+    denoiser_target_modules = {"transformer": ["to_qkv_mlp_proj", "to_k"]}
+
+
+class TestFlux2PipelineLoRAMemory(Flux2PipelineTesterConfig, LoraMemoryTesterMixin):
+    """LoRA x memory-optimization tests for the Flux2 pipeline."""
+
+    denoiser_target_modules = {"transformer": ["to_qkv_mlp_proj", "to_k"]}
